[shared] fix compiler warnings

+++ src/headers/string_op.h
@@ -24,7 +24,7 @@
 void os_trimcrlf(char *str);
 
 /* Similiar to Perl's substr() function */
-int os_substr(char *dest, const char *src, int position, int length);
+int os_substr(char *dest, const char *src, size_t position, size_t length);
 
 /* Remove a character from a string */
 char *os_strip_char(char *source, char remove);
